+++ .vs/GameClient/xs/UserPrefs.xml
@@ -1,13 +1,14 @@
 ﻿<Properties StartupConfiguration="{70B263AF-5707-DA2D-B24B-B59941E3F798}|">
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/src/Game/UDP_ClientController.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/src/Game/Client.cs">
     <Files>
-      <File FileName="Assets/src/Game/TCP_ClientController.cs" Line="12" Column="34" />
-      <File FileName="Assets/src/Game/UDP_ClientController.cs" Line="114" Column="39" />
+      <File FileName="Assets/src/Game/TCP_ClientController.cs" Line="103" Column="1" />
+      <File FileName="Assets/src/Game/UDP_ClientController.cs" Line="151" Column="1" />
       <File FileName="Assets/src/Game/PlayerController.cs" Line="39" Column="28" />
-      <File FileName="Assets/src/Library/BaseWeapon.cs" Line="1" Column="1" />
+      <File FileName="Assets/src/Library/BaseWeapon.cs" Line="41" Column="1" />
       <File FileName="Assets/src/Constant.cs" />
-      <File FileName="Assets/src/Game/Client.cs" Line="142" Column="18" />
+      <File FileName="Assets/src/Game/Client.cs" Line="363" Column="2" />
       <File FileName="Assets/src/Library/Convert.cs" />
+      <File FileName="Assets/src/Game/通信/UserBodyData.cs" Line="2" Column="14" />
     </Files>
     <Pads>
       <Pad Id="ProjectPad">
@@ -17,7 +18,8 @@
               <Node name="src" expanded="True">
                 <Node name="Game" expanded="True">
                   <Node name="Bom" expanded="True" />
-                  <Node name="UDP_ClientController.cs" selected="True" />
+                  <Node name="通信" expanded="True" />
+                  <Node name="Client.cs" selected="True" />
                 </Node>
                 <Node name="Library" expanded="True">
                   <Node name="Animator" expanded="True" />
@@ -34,7 +36,7 @@
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.PinnedWatches />
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.ItemProperties.Assembly-CSharp PreferredExecutionTarget="252c341b-3264-4d13-a498-84e4464da2ed" />
+  <MonoDevelop.Ide.ItemProperties.Assembly-CSharp PreferredExecutionTarget="1724fc89-fe8b-43e8-bb48-f2227bc454c1" />
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
     <BreakpointStore />
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
