+++ Assets/src/Game/Communication.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b602d7bb698e46cb8c280bcc74339bf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
