+++ Assets/src/Game/Communication/UDP_ClientController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b243fbb2cb37a48dbbfb1e76ad74fe10
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
