+++ Assets/src/Game/Communication/UserBodyData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50f4b785faab046bbae2487a0c00e341
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
