--- Assets/src/Game/TCP_ClientController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 526f67aad3a4e45a6b83cf4bec00a32b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
