+++ Library/CurrentLayout.dwlt
@@ -111,10 +111,10 @@ MonoBehaviour:
     y: 30
     width: 1680
     height: 904
-  m_MinSize: {x: 683, y: 348}
-  m_MaxSize: {x: 12008, y: 8048}
+  m_MinSize: {x: 683, y: 498}
+  m_MaxSize: {x: 14004, y: 14048}
   vertical: 0
-  controlID: 35
+  controlID: 32
 --- !u!114 &6
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -136,10 +136,10 @@ MonoBehaviour:
     y: 0
     width: 1339
     height: 904
-  m_MinSize: {x: 406, y: 348}
-  m_MaxSize: {x: 8006, y: 8048}
+  m_MinSize: {x: 406, y: 498}
+  m_MaxSize: {x: 10002, y: 14048}
   vertical: 1
-  controlID: 12
+  controlID: 33
 --- !u!114 &7
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -160,11 +160,11 @@ MonoBehaviour:
     x: 0
     y: 0
     width: 1339
-    height: 631
+    height: 630
   m_MinSize: {x: 406, y: 224}
   m_MaxSize: {x: 8006, y: 4024}
   vertical: 0
-  controlID: 13
+  controlID: 34
 --- !u!114 &8
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -183,7 +183,7 @@ MonoBehaviour:
     x: 0
     y: 0
     width: 306
-    height: 631
+    height: 630
   m_MinSize: {x: 202, y: 224}
   m_MaxSize: {x: 4002, y: 4024}
   m_ActualView: {fileID: 13}
@@ -209,7 +209,7 @@ MonoBehaviour:
     x: 306
     y: 0
     width: 1033
-    height: 631
+    height: 630
   m_MinSize: {x: 204, y: 224}
   m_MaxSize: {x: 4004, y: 4024}
   m_ActualView: {fileID: 12}
@@ -236,17 +236,17 @@ MonoBehaviour:
   m_Position:
     serializedVersion: 2
     x: 0
-    y: 631
+    y: 630
     width: 1339
-    height: 273
-  m_MinSize: {x: 102, y: 124}
-  m_MaxSize: {x: 4002, y: 4024}
-  m_ActualView: {fileID: 18}
+    height: 274
+  m_MinSize: {x: 232, y: 274}
+  m_MaxSize: {x: 10002, y: 10024}
+  m_ActualView: {fileID: 17}
   m_Panes:
   - {fileID: 17}
   - {fileID: 18}
-  m_Selected: 1
-  m_LastSelected: 0
+  m_Selected: 0
+  m_LastSelected: 1
 --- !u!114 &11
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -296,7 +296,7 @@ MonoBehaviour:
     x: 306
     y: 75
     width: 1029
-    height: 611
+    height: 610
   m_PersistentViewDataDictionary: {fileID: 0}
   m_WindowGUID: 49059b875290246fe849ec38390054cd
   m_SceneLighting: 1
@@ -304,9 +304,9 @@ MonoBehaviour:
   m_isRotationLocked: 0
   m_AudioPlay: 0
   m_Position:
-    m_Target: {x: -0.014478922, y: 1.3674505, z: 1.0265589}
+    m_Target: {x: 514.5, y: 190.5, z: 0}
     speed: 2
-    m_Value: {x: -0.014478922, y: 1.3674505, z: 1.0265589}
+    m_Value: {x: 514.5, y: 190.5, z: 0}
   m_RenderMode: 0
   m_CameraMode:
     drawMode: 0
@@ -338,9 +338,9 @@ MonoBehaviour:
     speed: 2
     m_Value: {x: 0.08717229, y: -0.89959055, z: 0.21045254, w: 0.3726226}
   m_Size:
-    m_Target: 8.16178
+    m_Target: 1053.1464
     speed: 2
-    m_Value: 8.16178
+    m_Value: 1053.1464
   m_Ortho:
     m_Target: 0
     speed: 2
@@ -376,14 +376,14 @@ MonoBehaviour:
     x: 0
     y: 75
     width: 304
-    height: 611
+    height: 610
   m_PersistentViewDataDictionary: {fileID: 0}
   m_SceneHierarchy:
     m_TreeViewState:
       scrollPos: {x: 0, y: 0}
-      m_SelectedIDs: 
+      m_SelectedIDs: 4c3f0000
       m_LastClickedID: 0
-      m_ExpandedIDs: 1edeffffccfbffff
+      m_ExpandedIDs: eadbffffccfbffffd03e0000
       m_RenameOverlay:
         m_UserAcceptedRename: 0
         m_Name: 
@@ -431,7 +431,7 @@ MonoBehaviour:
     x: 306
     y: 75
     width: 1029
-    height: 611
+    height: 610
   m_PersistentViewDataDictionary: {fileID: 0}
   m_MaximizeOnPlay: 0
   m_Gizmos: 0
@@ -465,9 +465,9 @@ MonoBehaviour:
       x: 0
       y: 17
       width: 1029
-      height: 594
+      height: 593
     m_Scale: {x: 2, y: 2}
-    m_Translation: {x: 514.5, y: 297}
+    m_Translation: {x: 514.5, y: 296.5}
     m_MarginLeft: 0
     m_MarginRight: 0
     m_MarginTop: 0
@@ -475,14 +475,14 @@ MonoBehaviour:
     m_LastShownAreaInsideMargins:
       serializedVersion: 2
       x: -257.25
-      y: -148.5
+      y: -148.25
       width: 514.5
-      height: 297
+      height: 296.5
     m_MinimalGUI: 1
   m_defaultScale: 2
   m_TargetTexture: {fileID: 0}
   m_CurrentColorSpace: 0
-  m_LastWindowPixelSize: {x: 2058, y: 1222}
+  m_LastWindowPixelSize: {x: 2058, y: 1220}
   m_ClearInEditMode: 1
   m_NoCameraWarning: 1
   m_LowResolutionForAspectRatios: 01000000000100000100
@@ -599,7 +599,7 @@ MonoBehaviour:
   m_Pos:
     serializedVersion: 2
     x: 0
-    y: 707
+    y: 705
     width: 1337
     height: 254
   m_PersistentViewDataDictionary: {fileID: 0}
@@ -615,19 +615,19 @@ MonoBehaviour:
     m_ShowAllHits: 0
     m_SearchArea: 1
     m_Folders:
-    - Assets/Scenes
+    - Assets/src/Game
   m_ViewMode: 1
   m_StartGridSize: 64
   m_LastFolders:
-  - Assets/Scenes
+  - Assets/src/Game
   m_LastFoldersGridSize: -1
   m_LastProjectPath: /Users/oogakiakiya/Desktop/GameClient
   m_LockTracker:
     m_IsLocked: 0
   m_FolderTreeState:
-    scrollPos: {x: 0, y: 627}
-    m_SelectedIDs: ce2c0000
-    m_LastClickedID: 11470
+    scrollPos: {x: 0, y: 692}
+    m_SelectedIDs: 742c0000
+    m_LastClickedID: 11380
     m_ExpandedIDs: 00000000ca270000602800006a28000096280000762900008e290000d42a00002c2b00002e2b0000d82b0000142c0000742c0000a22c0000fa2d00002e2e000000ca9a3b
     m_RenameOverlay:
       m_UserAcceptedRename: 0
@@ -683,7 +683,7 @@ MonoBehaviour:
   m_ListAreaState:
     m_SelectedInstanceIDs: 
     m_LastClickedInstanceID: 0
-    m_HadKeyboardFocusLastEvent: 0
+    m_HadKeyboardFocusLastEvent: 1
     m_ExpandedInstanceIDs: c6230000021000006a370000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
@@ -733,7 +733,7 @@ MonoBehaviour:
   m_Pos:
     serializedVersion: 2
     x: 0
-    y: 706
+    y: 707
     width: 1337
     height: 253
   m_PersistentViewDataDictionary: {fileID: 0}
