+++ Library/StateCache/SceneView/49059b-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-0.014478921890258789,"y":1.367450475692749,"z":1.0265588760375977},"rotation":{"x":0.08717229217290878,"y":-0.8995905518531799,"z":0.21045254170894624,"w":0.37262260913848879},"size":8.16178035736084,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":514.5,"y":190.5,"z":0.0},"rotation":{"x":0.08717229217290878,"y":-0.8995905518531799,"z":0.21045254170894624,"w":0.37262260913848879},"size":1053.1463623046875,"orthographic":false}
